跳转执行顺序和增加程序集文件夹管理健壮性

+++ HostedServices/ClearDynamicEndpointAssmeblyHostedService.cs
@@ -2,6 +2,7 @@
 using Microsoft.Extensions.Hosting;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -13,9 +14,13 @@ namespace DynamicEndpoint.HostedServices
         public Task StartAsync(CancellationToken cancellationToken)
         {
             // 清理动态路由程序集缓存
-            string[] files = Directory.GetFiles(new ClassTypeModel().AssemblyDir);
-            foreach (string item in files)
-                File.Delete(item);
+            string path = new ClassTypeModel().AssemblyDir;
+            if (Directory.Exists(path))
+            {
+                string[] files = Directory.GetFiles(new ClassTypeModel().AssemblyDir);
+                foreach (string item in files)
+                    File.Delete(item);
+            }
             return Task.CompletedTask;
         }
         public Task StopAsync(CancellationToken cancellationToken)
